+++ app/Http/Controllers/ThuctapsinhController.php
@@ -47,14 +47,14 @@ class ThuctapsinhController extends Controller
      */
     public function create()
     {
-        // if (Gate::allows('get-quantrivien')) {
+        if (Gate::allows('get-quantrivien')) {
         $title = "Thêm thực tập sinh";
         $get_dotthuctap = dot_thuctap::all();
         $get_chucvu = chucvu::all();
-        // } else {
-        //     return abort(403);
-        // }
-        return view('qtv.thuctapsinh.add', compact('title','get_dotthuctap','get_chucvu'));
+        } else {
+            return back();
+        }
+        return view('thuctapsinh.add', compact('title','get_dotthuctap','get_chucvu'));
     }
 
     /**
@@ -126,7 +126,7 @@ class ThuctapsinhController extends Controller
     public function show($id)
     {
         $title = "Cập nhật thực tập sinh";
-        return view('qtv.thuctapsinh.edit', compact('title'));
+        return view('thuctapsinh.edit', compact('title'));
     }
 
     /**
@@ -137,15 +137,15 @@ class ThuctapsinhController extends Controller
      */
     public function edit($id_sv)
     {
-        // if (Gate::allows('get-quantrivien')) {
+        if (Gate::allows('get-quanli')) {
         $title = "Cập nhật thực tập sinh";
         $get_dotthuctap = dot_thuctap::all();
         $get_chucvu = chucvu::all();
         $t= User::find($id_sv);
-        // } else {
-        //     return abort(403);
-        // }
-        return view('qtv.thuctapsinh.edit',compact('t','title','get_chucvu','get_dotthuctap'));
+        } else {
+            return back();
+        }
+        return view('thuctapsinh.edit',compact('t','title','get_chucvu','get_dotthuctap'));
     }
 
     /**
@@ -213,12 +213,12 @@ class ThuctapsinhController extends Controller
      */
     public function destroy($id_sv)
     {
-        // if (Gate::allows('get-quantrivien')) {
+        if (Gate::allows('get-quanli')) {
         $t= User::find($id_sv);
         $t->delete();
-        // } else {
-        //     return abort(403);
-        // }
+        } else {
+            return back();
+        }
         return redirect()->back()->with(['success' => 'Xóa thành công !']);
     }
